Avoid segmentation fault in case that no nodes are present.
A segmentation fault occured in GModelSpectralNodes::update_flux_cache when no nodes are present. The reason is that m_energies.size() returns an unsigned integer, and m_energies.size()-1 is not -1 for no nodes but 18446744073709551615, hence the loop was entered despite the fact that no nodes existed.

No, m_energies.size() is converted to int first, and then nodes-1 should give indeed -1 if nodes=0.