Put the test suite test case before the individual test cases that are done within a test suite.
GTestSuite::run starts with creating a test case that collects the total time spent in a test and that also collects any failures of the individual test cases. Before, this test case was appended AFTER running the individual test cases, which led to an XML file out of order. Now the general test case is appended BEFORE running the test cases.