Reconsider calibration database handling.
Instead of passing a string providing the calibration database path, the GCTAObservation and GCTAResponse classes now take a GCaldb object. This allows for a more flexible handling and selection of response functions through passing the corresponding calibration database object.

The rmffile member and method has become obsolete in GCTAResponse, so it was deleted. Now everything is handled by the GCTAEdisp class.