Change GResponse::has_edisp() to GResponse::use_edisp().
This change was prompted by the addition of the ctobssim option
to apply energy dispersion. Even if an instrument has an energy
dispersion, it is useful to be able to apply it or not.

So the meaning now is roughly this:
use_edisp = (has_edisp && apply_edisp)
although the apply_edisp flag has so far been only added to
GCTAResponse, use_edisp = False.

GResponse::has_tdisp() was also renamed to GResponse::use_tdisp()
for consistency.