Implement energy intergation in GModelSky.
The energy integration has been implemented in the GModelSky::integrate_energy method. For that purpose, the GModelSky::edisp_kern class has been added to GModelSky that provides the kernel for the integration.

The GResponse::src_ebounds method has been added to determine the true energy boundaries for a given observed event energy. The method is virtual in GResponse and always throws a "feature not implemented" exception. The method has then also been implemented in GCTAResponse to provide the true energy boundaries.