Don't use analytical gradients in case that energy dispersion is available.
If energy dispersion is available, the analytical gradients are wrong as they are not integrated over the energy dispersion PDF. We thus switch to numerical gradients in this case, which works pretty easy.

This is somehow a kluge, but not sure that integrating the analytical gradients numerically would at the end be faster.