Code integration review.
The code has been reviewed following integration. The code has be reformatted to fit the gammalib coding conventions.

In GCTAInstDir, the detx() and dety() methods have been declared inline in the header file.

For GCTAModelBackground, the code has been modified to work also properly when the number of energy bins is specified in the constructor. The default arguments for number of sky pixels and energy bins has been set to 0, as this is sufficient for the internal check (the -999 is somehow arbitrary, 0 seems to be more "well defined").

In GCTAPointing, the instdir() and skydir() methods do no longer return references. Returning references was not okay here as the objects were allocated locally within the methods, and the objects run out of scope when leaving the methods. The references are thus no longer valid.

In GCTAResponseTable, the new index and weight members have been added to init_members() and copy_members(). Furthermore, methods have been added that allow to retrieve the axis and parameter units.