Implement handling of instrumental background model via an XML file.
The instrumental background model may now be specified using an XML file. An example file is provided by 'cta_modelbg_obs.xml'. Note that one component is needed per observation in this XML file, and each component needs to be associated with the corresponding ID. Hence so far there is no mechanism to tie instrumental background models for different runs.