Take into account the spatial model value in Monte Carlo simulations (#1135).
The norm() method has been added to the abstract GModelSpatial interface that returns the model normalization within a circular region. This is used in the GModelSky::mc method to correctly reflect the normalization specified in the XML file. The method is also used to flag sources outside the simulation cone (applies so for the point sources, radial source and elliptical sources).