Throw exceptions if response information is missing.
So far the GCTAResponse class simply returned a zero response in case that no response information was loaded. This is not very informative for a user, who may debug for a long time before recognizing that he forgot to load response information.

Now, exceptions are thrown if response information is missing, telling the user that he has to load the response properly. This should be much saver.