Implement internal background model rebinning for MC sampling.
The actual background models provided with the IRF are pretty coarses, leading to measurable differences between MC simulated and modelled event distributions. To reduce the discrepancy, an internal rebinning of the background model in the spatial and energy dimensions has been implemented. Rebinning parameters are defined as maximum allowed spatial and spectral bin sizes. Parameters have been fixed to values that give satisfactory results for a specific test case (Arp220).