Add automatic type casting in Python interface for GObservation* return values.
Automatic type casting was implemented in the GObservations.i interface using the runtime type casting information that is maintained by the SWIG interface. The function relies on a newly introduced type() method in GObservation that provides as a string the class type.

The former GCTAObservation type casting constructor has been removed as it now should be obsolete.