Implement CSV table constructor, set methods and a save method.
A constructor has been added to create a table with a specific number of rows and columns. This constructor can be used for table creation.

Set methods string(), real() and integer() have been added to set the table cells to specific values.

A save() method has been implemented that will save the CSV table into an ASCII file.