Improve elliptical Irf and Npred computation.
The binned analysis looks okay, but unbinned shows a strange behaviour for the semiminor axis parameter which results in a strong dependency of Npred on the semiminor axis value. This seems to prevent a nice convergence in unbinned mode.

There is also an open issue to be solved related to the possibility that the minor axis is larger than the major axis. The actual code assumes that minor axis < major axis. If this is not fulfilled, the omega arc computations will be wrong.