Do not set XML filename upon loading of response components.
The setting of XML filenames upon loading of response components leads to an automatic saving of all response file paths in an XML file. This is very constraining and should be avoided. We therefore only save XML filenames that have been loaded, so that an exact copy of the input file can be written on output.