Add flux() methods.
The flux() methods return the correct flux in a pixel considering that a bilinear interpolation is performed in the map. Flux is defined as intensity times solid angle.

This works so far only properly for WCS projections. For HealPix maps, only the pixel intensity times the solid angle is returned.