Use precision in flux integration instead of a fixed number of iterations.
Switching to a fixed number of iterations led to an inaccuracy of the flux determination that affected considerably the pull distributions. In short, too many events have been simulated. I thus switched back to using the integration precision instead of using a fixed number of iterations, which should be okay for flux computation (flux computation is not involved in likelihood computations).