Checks users pfiles directory only if the PFILES environment variable is not set.
This handles the special case that the PFILES environment variable is set, yet the parameter file does not exist. Before, in that case the users pfiles directory was searched for the parameter file. Now, if the parameter file is missing, it will be copied in from the syspfiles directory, which is the expected behaviour.