Set GModelSpatialEllipticalGauss::theta_max() value to 2.5 times the maximum of semimajor and semiminor axis.
Before the limitation was 5 times that values and this led to convergence problems when fitting the model. Limiting the model to 2.5 times the larger axis seems to avoid the problem. Note that this is certainly not a perfect fix of the problem, but it helps to make the working model for now. Solves at least #1561.