Add Fermi LAT time reference kluge for source maps.
The Fermi LAT Science Tools do not write correctly the time reference for the Good Time Intervals into the source maps generated by gtsrcmaps. This prevents the LAT interface from reading in correctly the times. A kluge was added to add the correct LAT time reference in case that it is missing in the FITS file.