Add weights to CTA event cube (#1789)
A weighting sky map has been added to the CTA event cube that allows to correctly take into account data space selections. This is in particular needed for stacking event lists with different energy thresholds in an event cube, but can also be used for combining event lists with different ROIs.