Add model parameters to spectral registry (#1824) and revise spectral model names (#1821)
The GModelSpectralRegistry::alloc() method now also checks for the existence of model parameters and allocates the appropriate model from the registry that is found in the XML file base on model type and parameter names. This allows now to have the same model type but different parameter names in the registry.

The spectral model classes have been adapted to the new logic, and models have been added to the registry for all parameter combinations that may exist.

The unit tests and documentation have been updated.