Save application parameters only if they have been loaded
If an application is constructed from the command line arguments with the --help option the application parameters are not loaded, and the destructor saved before the empty parameter container to disk, which effectively emptied the parameter file. Now the desctructor only save the parameters if they have been loaded before.