Initialise WCS in base class method to avoid problem with double initalisation through wcs_set (#1851).
Formerly the wcs_set() method was called in the derived classes, which led problems with the GWcsGLS class that derives from GWcsSFL. Since GWcsSFL class wcs_set() with parameters that are intended for GWcsGLS, an exception has occured in the initial class. This problem is circumvented by placing the wcs_set call in the base class so that it will only be called once by any of the derived classes.