Change order of Python module imports
The order of Python module imports seems to impact the type casting capabilities. The import of the "model" module was moved down since when "GSkyRegionCircle" is used as an argument in the "model" module there is no casting to proxy classes anymore, and the Python interface does not work properly for sky regions. I moved a couple of other modules up since they may exhibit similar problems (#2961).