Use rate_ebin() method for CTA IRF background model evaluation (#3472)
The GCTAModelIrfBackground::eval() method uses now the GCTABackground3D::rate_ebin() integrator method instead of the GCTABackground3D::operator() operator so that the background model evaluation is no longer sensitive to the number of energy bins.

The GCTACubeBackground::fill() method was simplified, removing the power law integration which was not accurate enough, and using directly the GCTAModelIrfBackground::eval() method which is now exact.