Handle transparently the fitting of a model without any parameters
The fitting of a model without any parameters did not work following the introduction of the vector-wise response evaluation due to a constraint in the GObservations::likelihood::eval() method that imposed at minima one parameter. This was a relict of the old matrix interface, where the creation of an empty matrix was not allowed. The constraint was removed, which makes the code more general.