Finalise implementation of pulsar analysis support (#3973)
The implementation of pulsar analysis support was now fully implemented in GCOMDri (and specifically the compute_dre(), compute_drg() and compute_drx() methods). A pulsar phase selection is performed if pulsar ephemerides are present in GCOMSelection and if phase intervals are specified. If a pulsar phase selection is to be performed without the specification of BVC files an exception is thrown.

A PHASECOR keyword was added to the DRE header that gives the length of the phase interval that was used for the phase selection. Multiplying the IRF with this factor leads to a phase-selection corrected flux in the analysis.